rearranged files as per docs/development-setup.md fixed broken imports
