+++ src/components/core/Card.tsx
@@ -1,7 +1,7 @@
 import clsx from 'clsx';
 import React, { FC } from 'react';
 import { OverlayTrigger, Tooltip } from 'react-bootstrap';
-import { Breakpoints, Theme } from '../../context/ThemeProvider';
+import { Breakpoints, Theme } from '../../themes';
 import { createStyles } from '../../hooks/useTheme';
 import { agnosticFunctor } from '../../utils/functor';
 import Tag, { TagProps } from './Tag';
