+++ src/themes/index.ts
@@ -1,0 +1,2 @@
+export * from './theme';
+export * from './defaults';
