+++ src/themes/material/index.ts
@@ -1,0 +1,3 @@
+export * from './palette';
+export * from './typography';
+export * from './defaults';
